+++ CHANGELOG.md
@@ -1,5 +1,9 @@
-# Changelog
+# Change Log
 
 All notable changes to the "mib-codepilot-vscode" extension will be documented in this file.
 
+Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+
 ## [Unreleased]
+
+- Initial release